requirements.txt
# Core ML and NLP libraries
torch>=1.9.0
torchvision>=0.10.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
requests>=2.25.0
jinja2>=3.0.0
keybert>=0.8.0
spacy>=3.4.0
numpy>=1.21.0
tiktoken>=0.5.0

# CLIP dependencies
ftfy>=6.0.0
regex>=2021.0.0
tqdm>=4.62.0

# Image processing
pillow>=8.3.0
opencv-python>=4.5.0

# PDF processing
pdf2image>=1.16.3
pdfplumber>=0.11.0
PyMuPDF>=1.24.0

# Text processing and evaluation
nltk>=3.6.0
rouge>=1.0.0
pytesseract>=0.3.8

# Web framework
streamlit>=1.28.0

# Document processing
python-docx>=0.8.11

# Data processing and visualization
pandas>=1.3.0
plotly>=5.0.0

# Environment variables
python-dotenv>=0.19.0

# CLIP installation (requires git)
git+https://github.com/openai/CLIP.git